маленький рефактор
